+++ g2engine/g2engine.go
@@ -24,13 +24,13 @@ import (
 	"time"
 	"unsafe"
 
+	"github.com/senzing-garage/g2-sdk-go/g2api"
+	g2engineapi "github.com/senzing-garage/g2-sdk-go/g2engine"
+	"github.com/senzing-garage/g2-sdk-go/g2error"
 	"github.com/senzing-garage/go-logging/logging"
 	"github.com/senzing-garage/go-observing/notifier"
 	"github.com/senzing-garage/go-observing/observer"
 	"github.com/senzing-garage/go-observing/subject"
-	"github.com/senzing/g2-sdk-go/g2api"
-	g2engineapi "github.com/senzing/g2-sdk-go/g2engine"
-	"github.com/senzing/g2-sdk-go/g2error"
 )
 
 // ----------------------------------------------------------------------------