+++ g2engine/g2engine_test.go
@@ -13,17 +13,17 @@ import (
 	"time"
 
 	truncator "github.com/aquilax/truncate"
+	"github.com/senzing-garage/g2-sdk-go-base/g2config"
+	"github.com/senzing-garage/g2-sdk-go-base/g2configmgr"
+	"github.com/senzing-garage/g2-sdk-go/g2api"
+	g2engineapi "github.com/senzing-garage/g2-sdk-go/g2engine"
+	"github.com/senzing-garage/g2-sdk-go/g2error"
 	futil "github.com/senzing-garage/go-common/fileutil"
 	"github.com/senzing-garage/go-common/g2engineconfigurationjson"
 	"github.com/senzing-garage/go-common/record"
 	"github.com/senzing-garage/go-common/testfixtures"
 	"github.com/senzing-garage/go-common/truthset"
 	"github.com/senzing-garage/go-logging/logging"
-	"github.com/senzing/g2-sdk-go-base/g2config"
-	"github.com/senzing/g2-sdk-go-base/g2configmgr"
-	"github.com/senzing/g2-sdk-go/g2api"
-	g2engineapi "github.com/senzing/g2-sdk-go/g2engine"
-	"github.com/senzing/g2-sdk-go/g2error"
 	"github.com/stretchr/testify/assert"
 )