+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed in 0.5.0
 
-- Renamed module to github.com/senzing-garage/g2-sdk-go-base
+- Renamed module to `github.com/senzing-garage/g2-sdk-go-base`
 - Refactor to [template-go](https://github.com/senzing-garage/template-go)
 - Update dependencies
   - google.golang.org/grpc v1.61.0